extract ExtendableHTMLElement to its own module
